project/requirements.txt
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
Flask==3.1.2
flask-cors==6.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Naked==0.1.32
pycparser==2.23
pycryptodome==3.23.0
pymongo==4.15.3
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
shellescape==3.8.1
urllib3==2.5.0
Werkzeug==3.1.3
